extract as file
